[Refactor] budgetcontent로 패키징
