+++ app/src/main/res/layout/activity_budget_content.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.budget.BudgetContentActivity">
+    tools:context=".ui.budget.budgetcontent.BudgetContentActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/budget_content_toolbar"
